add a condition

+++ JavPy/tests/test_clients/test_telegram.py
@@ -1,6 +1,7 @@
 import shutil
 import os
-shutil.rmtree(os.path.join(os.path.expanduser("~"), ".JavPy"), ignore_errors=True)
+if "GITHUB_WORKFLOW" in os.environ:
+    shutil.rmtree(os.path.join(os.path.expanduser("~"), ".JavPy"), ignore_errors=True)
 
 import requests
 from JavPy.app.tgbot.server import (
